refactor(taskctl): split pulse.ts and tool.ts into focused modules under 500 lines (#224)
